requirements.txt
torch==2.0.1;sys_platform == 'darwin'
torch;sys_platform != 'darwin'
faiss-cpu
scipy
scikit-learn
numpy
transformers
